Created IrcConnection and IIrcConnection. Moved the contents of Bot in to IrcConnection. Removed Bot as it was no longer used.
