--- src/SpikeCore/SpikeCore/IBot.cs
@@ -1,7 +1,0 @@
-﻿namespace SpikeCore
-{
-    public interface IBot
-    {
-        // This is an empty marker interface for now.
-    }
-}
