+++ src/SpikeCore/SpikeCore/Irc/IIrcConnection.cs
@@ -1,0 +1,4 @@
+﻿namespace SpikeCore.Irc
+{
+    public interface IIrcConnection { }
+}
